Move admin client into it's own factory as requested during the PR
